further timescaledb configuration exploration; complete questdb data layer method set; configure directory structure with minimal api usage in mind; add serilog; add request logging filter

+++ backend-dotnet/Models/MarketBar.cs
@@ -4,10 +4,10 @@ public class MarketBar
 {
     public DateTime Timestamp { get; set; }
     public string Symbol { get; set; } = string.Empty;
-    public decimal Open { get; set; }
-    public decimal High { get; set; }
-    public decimal Low { get; set; }
-    public decimal Close { get; set; }
+    public float Open { get; set; }
+    public float High { get; set; }
+    public float Low { get; set; }
+    public float Close { get; set; }
     public long Volume { get; set; }
     public string Timeframe { get; set; } = string.Empty;
 }
